refactor: adjust file tree and other logic according to pr review
