+++ src/app/[locale]/layout.tsx
@@ -1,6 +1,6 @@
 import { SpeedInsights } from '@vercel/speed-insights/next'
 import { capitalize } from 'lodash-es'
-import type { Metadata, Viewport } from 'next'
+import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import NextTopLoader from 'nextjs-toploader'
 
@@ -19,6 +19,8 @@ import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
 import { fontClassName } from '@/utils/web/fonts'
 
+export { viewport } from '@/utils/server/metadataUtils'
+
 // we want dynamicParams to be false for this top level layout, but we also want to ensure that subpages can have dynamic params
 // Next.js doesn't allow this so we allow dynamic params in the config here, and then trigger a notFound in the layout if one is passed
 // export const dynamicParams = false
@@ -34,31 +36,11 @@ const title = `${
 const description = `Stand With Crypto Alliance is a non-profit organization dedicated to uniting global crypto advocates.`
 const ogImage = getOpenGraphImageUrl({ title: description })
 
-export const viewport: Viewport = {
-  viewportFit: 'cover',
-  width: 'device-width',
-  initialScale: 1,
-  maximumScale: 1,
-  themeColor: '#fff',
-}
-
 export const metadata: Metadata = {
   ...generateMetadataDetails({ description, title, ogImage }),
   title: {
     default: title,
     template: '%s | Stand With Crypto',
-  },
-  metadataBase: new URL('https://www.standwithcrypto.org'),
-  applicationName: 'Stand With Crypto',
-  icons: [
-    { url: '/logo/favicon-16x16.png', sizes: '16x16' },
-    { url: '/logo/favicon-32x32.png', sizes: '32x32' },
-  ],
-  // manifest: '/site.webmanifest', // LATER-TASK figure out why we get 401s when we uncomment this
-  appleWebApp: {
-    title: 'Stand With Crypto',
-    statusBarStyle: 'black-translucent',
-    startupImage: ['/logo/apple-touch-icon.png'],
   },
 }
