+++ src/app/embedded/layout.tsx
@@ -1,22 +1,17 @@
 import { SpeedInsights } from '@vercel/speed-insights/next'
 import { capitalize } from 'lodash-es'
-import type { Metadata, Viewport } from 'next'
+import { Metadata } from 'next'
 
 import { OverrideGlobalLocalStorage } from '@/components/app/overrideGlobalLocalStorage'
 import { FullHeight } from '@/components/ui/fullHeight'
-import { ORDERED_SUPPORTED_LOCALES, SupportedLocale } from '@/intl/locales'
+import { SupportedLocale } from '@/intl/locales'
 import { PageProps } from '@/types'
 import { getOpenGraphImageUrl } from '@/utils/server/generateOpenGraphImageUrl'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
 import { NEXT_PUBLIC_ENVIRONMENT } from '@/utils/shared/sharedEnv'
 import { fontClassName } from '@/utils/web/fonts'
 
-// we want dynamicParams to be false for this top level layout, but we also want to ensure that subpages can have dynamic params
-// Next.js doesn't allow this so we allow dynamic params in the config here, and then trigger a notFound in the layout if one is passed
-// export const dynamicParams = false
-export async function generateStaticParams() {
-  return ORDERED_SUPPORTED_LOCALES.map(locale => ({ locale }))
-}
+export { viewport } from '@/utils/server/metadataUtils'
 
 const title = `${
   NEXT_PUBLIC_ENVIRONMENT === 'production'
@@ -26,31 +21,11 @@ const title = `${
 const description = `Stand With Crypto Alliance is a non-profit organization dedicated to uniting global crypto advocates.`
 const ogImage = getOpenGraphImageUrl({ title: description })
 
-export const viewport: Viewport = {
-  viewportFit: 'cover',
-  width: 'device-width',
-  initialScale: 1,
-  maximumScale: 1,
-  themeColor: '#fff',
-}
-
 export const metadata: Metadata = {
   ...generateMetadataDetails({ description, title, ogImage }),
   title: {
     default: title,
     template: '%s | Stand With Crypto',
-  },
-  metadataBase: new URL('https://www.standwithcrypto.org'),
-  applicationName: 'Stand With Crypto',
-  icons: [
-    { url: '/logo/favicon-16x16.png', sizes: '16x16' },
-    { url: '/logo/favicon-32x32.png', sizes: '32x32' },
-  ],
-  // manifest: '/site.webmanifest', // LATER-TASK figure out why we get 401s when we uncomment this
-  appleWebApp: {
-    title: 'Stand With Crypto',
-    statusBarStyle: 'black-translucent',
-    startupImage: ['/logo/apple-touch-icon.png'],
   },
 }
