+++ src/components/app/pageAdvocatesHeatmap/constants.ts
