+++ src/data/aggregations/getCountUsers.ts
@@ -12,7 +12,7 @@ export const getCountUsers = async () => {
   Our database in testing env is populated with way less info but we want the UI
   to look comparable to production so we mock the numbers
   */
-  return { count: count * 1011 }
+  return { count: count * 10111 }
 }
 
 export type CountUsers = Awaited<ReturnType<typeof getCountUsers>>
